rental_site/requirements.txt
Django==3.0
django-ckeditor==6.0.0
